Reworked service classes, not finished

+++ SAES-service/src/main/java/cz/muni/fi/pa165/service/SportActivityServiceImpl.java
@@ -1,13 +1,16 @@
 package cz.muni.fi.pa165.service;
 
+import cz.muni.fi.pa165.saes.dao.ActivityRecordDao;
+import cz.muni.fi.pa165.saes.dao.BurnedCaloriesDao;
 import cz.muni.fi.pa165.saes.dao.SportActivityDao;
+import cz.muni.fi.pa165.saes.entity.ActivityRecord;
 import cz.muni.fi.pa165.saes.entity.SportActivity;
-import cz.muni.fi.pa165.service.exceptions.SaesDataAccessException;
-import cz.muni.fi.pa165.service.exceptions.SaesServiceException;
-import java.util.List;
-import javax.inject.Inject;
+import cz.muni.fi.pa165.service.exceptions.EntityReferenceException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import javax.inject.Inject;
+import java.util.List;
 
 /**
  * Implementation of the {@link SportActivityService}.
@@ -21,55 +24,46 @@ public class SportActivityServiceImpl implements SportActivityService {
     @Inject
     private SportActivityDao sportActivityDao;
 
+    @Inject
+    private BurnedCaloriesDao burnedCaloriesDao;
+
+    @Inject
+    private ActivityRecordDao activityRecordDao;
+
     @Override
     public SportActivity create(SportActivity activity) {
-        try {
-            sportActivityDao.createSportActivity(activity);
-        } catch (Exception e) {
-            throw new SaesDataAccessException("Failed when creating sport activity. ", e);
-        }
+        sportActivityDao.createSportActivity(activity);
 
         return activity;
     }
 
     @Override
-    public void delete(SportActivity activity) {
-        try {
-            sportActivityDao.deleteSportActivity(activity);
-        } catch (Exception e) {
-            throw new SaesDataAccessException("Failed when deleting sport activity. ", e);
+    public void delete(Long id) throws EntityReferenceException {
+        SportActivity activity = sportActivityDao.findSportActivity(id);
+        if (activity != null) {
+            List<ActivityRecord> recordsBySportActivity = activityRecordDao.findRecordsBySportActivity(activity);
+            if (recordsBySportActivity.isEmpty()) {
+                sportActivityDao.deleteSportActivity(activity);
+                burnedCaloriesDao.deleteBySportActivity(activity);
+            } else {
+                throw new EntityReferenceException("Cannot delete sport activity, there are " +
+                        recordsBySportActivity.size() + " referenced to it. ");
+            }
         }
     }
 
     @Override
     public SportActivity findById(Long id) {
-        try {
-            return sportActivityDao.findSportActivity(id);
-        } catch (Exception e) {
-            throw new SaesDataAccessException("Failed when finding sport activity. ", e);
-        }
+        return sportActivityDao.findSportActivity(id);
     }
 
     @Override
     public List<SportActivity> findAll() {
-        try {
-            return sportActivityDao.findAll();
-        } catch (Exception e) {
-            throw new SaesDataAccessException("Failed when finding all sport activities. ", e);
-        }
+        return sportActivityDao.findAll();
     }
 
     @Override
-    public void changeName(SportActivity activity, String newName) {
-        if (activity.getName().equals(newName)) {
-            throw new SaesServiceException("Activity is already named as '" + newName + "'. ");
-        }
-        activity.setName(newName);
-
-        try {
-            sportActivityDao.updateSportActivity(activity);
-        } catch (Exception e) {
-            throw new SaesDataAccessException("Failed when changing name of sport activity. ", e);
-        }
+    public void update(SportActivity activity) {
+        sportActivityDao.updateSportActivity(activity);
     }
 }
